refactor: 멤버 서비스 수정(#51)
