+++ src/main/java/com/flab/tabling/global/exception/BusinessException.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 
 @Getter
 public class BusinessException extends RuntimeException {
-	private static final String DEFAULT_MESSAGE = "common business exception occurs";
 	private final ErrorCode errorCode;
 
 	public BusinessException(ErrorCode errorCode, String message) {
