+++ src/main/java/com/flab/tabling/member/facade/MemberFacade.java
@@ -9,6 +9,7 @@ import com.flab.tabling.global.service.SessionService;
 import com.flab.tabling.global.service.TwoWayCipherService;
 import com.flab.tabling.member.domain.Member;
 import com.flab.tabling.member.dto.MemberAddDto;
+import com.flab.tabling.member.service.MemberQueryService;
 import com.flab.tabling.member.service.MemberService;
 
 import jakarta.servlet.http.HttpSession;
@@ -20,13 +21,14 @@ public class MemberFacade {
 	private final OneWayCipherService oneWayCipherService;
 	private final TwoWayCipherService twoWayCipherService;
 	private final MemberService memberService;
+	private final MemberQueryService memberQueryService;
 	private final SessionService sessionService;
 
 	@Transactional
 	public MemberAddDto.Response add(MemberAddDto.Request memberRequestDto, HttpSession session) {
 		String encryptedPassword = oneWayCipherService.encrypt(memberRequestDto.getPassword());
 		String encryptedEmail = twoWayCipherService.encrypt(memberRequestDto.getEmail());
-		memberService.checkEmailDuplicated(encryptedEmail);
+		memberQueryService.checkEmailDuplicated(encryptedEmail);
 		sessionService.add(session, SessionConstant.MEMBER_NAME, memberRequestDto.getName());
 		Member member = memberService.add(memberRequestDto.getName(), encryptedEmail, encryptedPassword,
 			memberRequestDto.getRoleType());
